Fixes build hang caused by active process tree
The build (more accurately, the test run phase) was getting stuck due to grand-child processes lingering on. To repro:
  1. cd test\Microsoft.AspNetCore.Tests.Performance
  2. dotnet restore
  3. dotnet build
  4. dotnet test

At the end, we're stuck waiting for the child process. The process we're waiting for in this case is `BasicKestrel.exe`. When I go to task manager and kill `BasicKestrel.exe`, immediately `corehost.exe` ends.

The code that's causing this goes as follows. In WebPerformance.cs we're calling `Process.Start()` with a `ProcessStartInfo` describing the command `dotnet run` on a deployment of BasicKestrel.

When the line that reads `process.Kill();` executes, we're effectively killing `dotnet.exe`, but not its child processes. Our test process spawned a `dotnet.exe` which in turn spawned `BasicKestrel.exe`. Our process then kills `dotnet.exe` but `BasicKestrel.exe` lingers around. Our process, however, is magically entangled with `BasicKestrel.exe` as we're waiting for it to end before we can shut down.

This commit changes the behavior to use an extension method that kills an entire process tree as opposed to just the process at hand.

This fixes #101.